backend/requirements.txt
fastapi
uvicorn[standard]
motor                 # async MongoDB driver
python-jose[cryptography]   # JWT (or PyJWT)
pydantic
python-multipart
aiofiles
